requirements.txt
pytest==7.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pytest-order==1.1.0
pytest-cov==4.1.0
Flask==2.3.2
SQLAlchemy==2.0.15